QuestionMap

An R6 QuestionMap object holds the mapping for one question or demographic between the survey and population dataset.

Methods

Public methods

QuestionMap$new()

QuestionMap$print()

QuestionMap$name()

QuestionMap$col_names()

QuestionMap$values()

QuestionMap$clone()

Method new()

Create a new QuestionMap

Usage

QuestionMap$new(name, col_names, values_map)

Arguments

name

The name of the underlying construct. For example if the sample data uses the name age1 and the population data uses the name age2 then name could be age.

col_names

A character vector of length two giving the column names of the question in the sample data (first element) and the population data (second element).

values_map

A named list where the names correspond to the responses to the question in the sample data and the values correspond to the responses in the population data. If there is a meaningful ordering over the values, they should be listed in that order, either descending or ascending. See Examples.

Returns

A QuestionMap object that can be added to a SurveyMap object.

Method print()

Print a QuestionMap.

Usage

QuestionMap$print(...)

Arguments

...

Currently ignored.

Returns

The QuestionMap object, invisibly.

Method name()

Access name

Usage

QuestionMap$name()

Method col_names()

Access col_names

Usage

QuestionMap$col_names()

Method values()

Access values

Usage

QuestionMap$values()

Method clone()

The objects of this class are cloneable with this method.

Usage

QuestionMap$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

q_age <- QuestionMap$new(
  name = "age",
  col_names = c("age","age_group"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
print(q_age)

#> -------------- 
#> age 
#> age = age_group 
#> -------------- 
#> 18-25 = 18-35 
#> 26-35 = 18-35 
#> 36-45 = 36-55 
#> 46-55 = 36-55 
#> 56-65 = 56-65 
#> 66-75 = 66+ 
#> 76-90 = 66+ 

q_party_pref <- QuestionMap$new(
  name = "party_pref",
  col_names = c("vote_for","vote_pref"),
  values_map = list("Box Party" = "The BP", "The BP" = "The BP","The Circle Party" = "The Circles","The Circles" = "The Circles")
)
q_gender <- QuestionMap$new(
  name = "gender",
  col_names = c("gender", "gender"),
  values_map = list("male" = "m","female" = "f", "nonbinary" = "nb")
)