SurveyData objects

A SurveyData object represents a survey and its metadata. The survey itself is a data frame containing all data from the survey. The survey data object also includes the survey questions and responses (if left empty these will just be the column and factor level names). To enable weighted comparisons, survey weights and a survey design can be specified, with the survey design specified using survey package notation.

Methods

Public methods

SurveyData$new()

SurveyData$n_obs()

SurveyData$n_questions()

SurveyData$print()

SurveyData$add_survey_data_column()

SurveyData$survey_data()

SurveyData$questions()

SurveyData$responses()

SurveyData$weights()

SurveyData$design()

SurveyData$clone()

Method new()

Create a new SurveyData object using an existing data frame and other survey information. This method is used to create the objects for both the sample and the population data. If a population is approximated from a large survey (like the ACS or DHS), then the package will enable the creation of a weighted poststratification matrix. If the population is summarized as a poststratification matrix already, then set the weights as the size in each cell $N_j$. If the entire individual level population data is given, then weights should be omitted and will be automatically set to 1.

Usage

SurveyData$new(
  data,
  questions = list(),
  responses = list(),
  weights = numeric(),
  design = list(ids = ~1)
)

Arguments

data

A data frame containing the survey data.

questions, responses

Named lists containing the text of the survey questions and the allowed responses, respectively. The names must correspond to the names of variables in data. See Examples. If these aren't provided then they will be created internally using all factor, character, and binary variables in data.

weights

Optionally, the name of a variable in data containing survey weights.

design

Optionally, a named list of arguments (except weights and data) to pass to survey::svydesign() to specify the survey design.

Examples

# Population estimated from large survey
popn_obj1 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  ),
  weights = "wt" # use the wt column from approx_voters_popn data
)

# Population poststratification matrix already known
library(dplyr)
popn_ps <- approx_voters_popn %>%
  group_by(age_group,gender) %>%
  summarise(N_j = sum(wt))

popn_obj2 <- SurveyData$new(
  data = popn_ps,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(popn_ps$gender),
    age_group = levels(popn_ps$age_group)
  ),
  weights = "N_j"# use N_j column from popn_ps data
)

# Individual population data known:
# Pretend that approx_voters_popn is the full population
popn_obj3 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  )
)
popn_obj1
popn_obj2
popn_obj3

Method n_obs()

Number of observations in the survey data

Usage

SurveyData$n_obs()

Method n_questions()

Number of survey questions

Usage

SurveyData$n_questions()

Method print()

Print a summary of the survey data

Usage

SurveyData$print(...)

Arguments

...

Currently ignored.

Method add_survey_data_column()

Add a column to the sample data. This is primarily intended for internal use but may occasionally be useful.

Usage

SurveyData$add_survey_data_column(name, value)

Arguments

name, value

The name of the new variable (a string) and the vector of values to add to the data frame.

Method survey_data()

Access the data frame containing the sample data.

Usage

SurveyData$survey_data(key = TRUE)

Arguments

key

Should the .key column be included? This column just indicates the original order of the rows and is primarily intended for internal use.

Method questions()

Access the list of survey questions

Usage

SurveyData$questions()

Method responses()

Access the list of allowed survey responses

Usage

SurveyData$responses()

Method weights()

Access the survey weights

Usage

SurveyData$weights()

Method design()

Access the survey design

Usage

SurveyData$design()

Method clone()

The objects of this class are cloneable with this method.

Usage

SurveyData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

head(shape_survey)

#>     age gender     vote_for   highest_educ   state   y        wt
#> 1 56-65   male           BP   some college State D  no  79.90290
#> 2 46-55   male Circle Party 4-year college State B yes  89.12213
#> 3 36-45   male Circle Party     associates State D  no  92.90569
#> 4 18-25   male    Box Party    high school State B yes  80.78384
#> 5 56-65 female           BP    high school State A yes  89.85756
#> 6 56-65 female Circle Party 4-year college State B yes 117.04200

box_prefs <- SurveyData$new(
  data = shape_survey,
  questions = list(
    age = "Please identify your age group",
    gender = "Please select your gender",
    vote_for = "Which party did you vote for in the 2018 election?",
    y = "If today is the election day, would you vote for the Box Party?"
  ),
  responses = list(
    age = levels(shape_survey$age),
    gender = levels(shape_survey$gender),
    vote_for = levels(shape_survey$vote_for),
    y = c("no","yes")
  ),
  weights = "wt",
  design = list(ids =~1)
)
box_prefs$print()

#> Survey with 500 observations, 4 questions 
#> Independent Sampling design (with replacement) 
#> 
#> Column label: age 
#> Question: Please identify your age group 
#> Allowed answers: 18-25, 26-35, 36-45, 46-55, 56-65, 66-75, 76-90 
#> 
#> Column label: gender 
#> Question: Please select your gender 
#> Allowed answers: male, female, nonbinary 
#> 
#> Column label: vote_for 
#> Question: Which party did you vote for in the 2018 election? 
#> Allowed answers: Box Party, BP, Circle Party, CP 
#> 
#> Column label: y 
#> Question: If today is the election day, would you vote for the Box Party? 
#> Allowed answers: no, yes 

box_prefs$n_questions()

#> [1] 4

head(approx_voters_popn)

#>   age_group gender vote_pref        wt       education state
#> 1       66+      m        BP  6.007461 4-years college     A
#> 2     56-65      f        CP  8.465851     high school     A
#> 3     56-65      m        CP  4.098878 4-years college     C
#> 4       66+      m        BP  3.883250 4-years college     C
#> 5       66+      f        CP 16.179562    some college     B
#> 6     18-35      m        BP  3.851837     high school     D

popn_obj <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?",
    vote_pref = "Which party do you prefer to vote for?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group),
    vote_pref = levels(approx_voters_popn$vote_pref)
  ),
  weights = "wt"
)
popn_obj$print()

#> Survey with 5000 observations, 3 questions 
#> Independent Sampling design (with replacement) 
#> 
#> Column label: age_group 
#> Question: Which age group are you? 
#> Allowed answers: 18-35, 36-55, 56-65, 66+ 
#> 
#> Column label: gender 
#> Question: Gender? 
#> Allowed answers: m, f, nb 
#> 
#> Column label: vote_pref 
#> Question: Which party do you prefer to vote for? 
#> Allowed answers: BP, CP 

## ------------------------------------------------
## Method `SurveyData$new`
## ------------------------------------------------

# Population estimated from large survey
popn_obj1 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  ),
  weights = "wt" # use the wt column from approx_voters_popn data
)

# Population poststratification matrix already known
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

popn_ps <- approx_voters_popn %>%
  group_by(age_group,gender) %>%
  summarise(N_j = sum(wt))

#> `summarise()` has grouped output by 'age_group'. You can override using the `.groups` argument.

popn_obj2 <- SurveyData$new(
  data = popn_ps,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(popn_ps$gender),
    age_group = levels(popn_ps$age_group)
  ),
  weights = "N_j"# use N_j column from popn_ps data
)

# Individual population data known:
# Pretend that approx_voters_popn is the full population
popn_obj3 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  )
)
popn_obj1

#> Survey with 5000 observations, 2 questions 
#> Independent Sampling design (with replacement) 
#> 
#> Column label: age_group 
#> Question: Which age group are you? 
#> Allowed answers: 18-35, 36-55, 56-65, 66+ 
#> 
#> Column label: gender 
#> Question: Gender? 
#> Allowed answers: m, f, nb 

popn_obj2

#> Survey with 12 observations, 2 questions 
#> Independent Sampling design (with replacement) 
#> 
#> Column label: age_group 
#> Question: Which age group are you? 
#> Allowed answers: 18-35, 36-55, 56-65, 66+ 
#> 
#> Column label: gender 
#> Question: Gender? 
#> Allowed answers: m, f, nb 

popn_obj3

#> Survey with 5000 observations, 2 questions 
#> Independent Sampling design (with replacement) 
#> 
#> Column label: age_group 
#> Question: Which age group are you? 
#> Allowed answers: 18-35, 36-55, 56-65, 66+ 
#> 
#> Column label: gender 
#> Question: Gender? 
#> Allowed answers: m, f, nb