Simulated Survey Data for Shape World

This is a simulated survey of whether Shape World voters prefer the Box Party (BP) or the Circle Party (CP) in the 2020 election. The data also include several demographic variables and the party the respondent voted for in the previous election. We use this data throughout the documentation to demonstrate the usage of the package.

shape_survey

Format

A data frame with columns:

age

Age group

gender

Gender

vote_for

Party voted for in the 2016 election

highest_educ

Highest grade of education completed

state

State

y

Preference for BP over CP in 2020 election

wt

Raked and trimmed survey weight

Source

Created in ShapeWorldVoters.R

Examples

 shape_survey

#>       age    gender     vote_for   highest_educ   state   y        wt
#> 1   56-65      male           BP   some college State D  no  79.90290
#> 2   46-55      male Circle Party 4-year college State B yes  89.12213
#> 3   36-45      male Circle Party     associates State D  no  92.90569
#> 4   18-25      male    Box Party    high school State B yes  80.78384
#> 5   56-65    female           BP    high school State A yes  89.85756
#> 6   56-65    female Circle Party 4-year college State B yes 117.04200
#> 7   66-75    female           BP    high school State B  no 128.27293
#> 8   56-65    female           BP   some college State B yes 291.96747
#> 9   66-75    female           BP 4-year college State C yes  74.93614
#> 10  66-75      male    Box Party    high school State D yes  19.47353
#> 11  18-25      male Circle Party    high school State C  no 133.85761
#> 12  36-45      male Circle Party 4-year college State B yes  99.55316
#> 13  56-65      male Circle Party   some college State B yes 164.24949
#> 14  76-90      male           CP 4-year college State E yes  27.52989
#> 15  66-75    female    Box Party    high school State E yes  43.12209
#> 16  66-75      male           BP   some college State B  no 166.92312
#> 17  36-45      male Circle Party   some college State D yes 197.33458
#> 18  66-75      male           BP     associates State B yes  78.91989
#> 19  46-55      male           CP    high school State B yes  49.06783
#> 20  56-65    female           BP 4-year college State E yes  62.15251
#> 21  36-45      male           BP    high school State B yes  83.79017
#> 22  36-45      male           BP 4-year college State A yes  71.11638
#> 23  76-90      male    Box Party     associates State B  no  43.94217
#> 24  66-75      male           BP    high school State B yes  57.40206
#> 25  66-75      male           BP    high school State B yes  57.40206
#> 26  56-65      male Circle Party 4-year college State A yes  48.18186
#> 27  66-75      male           CP   some college State C yes  67.76242
#> 28  76-90      male Circle Party   some college State B  no 242.00590
#> 29  46-55      male    Box Party 4-year college State A yes  32.25274
#> 30  56-65      male Circle Party  post-graduate State B yes  43.53646
#> 31  26-35      male           BP 4-year college State D yes 130.22139
#> 32  56-65      male    Box Party 4-year college State B yes  21.72295
#> 33  66-75      male    Box Party   some college State D yes  51.22441
#> 34  56-65      male Circle Party   some college State A yes 149.75773
#> 35  46-55      male    Box Party   some college State B  no 106.48465
#> 36  56-65    female Circle Party no high school State D yes 147.85062
#> 37  66-75      male Circle Party 4-year college State A yes  62.16866
#> 38  36-45      male           CP    high school State D yes  35.02715
#> 39  36-45      male Circle Party 4-year college State B yes  99.55316
#> 40  56-65    female           BP 4-year college State A yes  83.40113
#> 41  56-65      male Circle Party    high school State B yes  56.52559
#> 42  66-75    female Circle Party   some college State B  no 303.54540
#> 43  36-45    female Circle Party    high school State A yes 222.68809
#> 44  56-65    female           CP   some college State B  no 186.59744
#> 45  46-55      male           CP   some college State E yes  95.59828
#> 46  56-65      male           BP    high school State B yes  44.56118
#> 47  18-25      male    Box Party 4-year college State A  no  68.58751
#> 48  56-65    female           BP 4-year college State A  no  83.40113
#> 49  76-90      male           BP    high school State B  no  64.24643
#> 50  66-75      male    Box Party 4-year college State C yes  18.39985
#> 51  66-75    female Circle Party 4-year college State A yes 139.36402
#> 52  36-45 nonbinary Circle Party     associates State A yes 244.48791
#> 53  46-55      male Circle Party     associates State B  no 133.12846
#> 54  56-65      male           BP  post-graduate State B yes  34.53010
#> 55  36-45      male           CP 4-year college State D yes  32.69338
#> 56  46-55      male           CP   some college State E yes  95.59828
#> 57  56-65    female           BP    high school State B yes  98.39429
#> 58  56-65    female           BP 4-year college State B  no  91.29599
#> 59  56-65      male           CP 4-year college State B yes  27.78147
#> 60  46-55      male           BP no high school State B yes 140.38468
#> 61  66-75    female    Box Party   some college State A  no 166.97311
#> 62  76-90    female           BP 4-year college State D yes  83.52402
#> 63  76-90    female Circle Party    high school State C yes 116.05616
#> 64  36-45      male           BP   some college State B yes 247.41847
#> 65  66-75      male Circle Party    high school State B yes  73.15311
#> 66  46-55    female Circle Party    high school State A  no 198.81667
#> 67  66-75      male Circle Party   some college State B yes 214.97076
#> 68  56-65    female    Box Party     associates State B  no  66.58369
#> 69  56-65      male           CP     associates State B yes  40.08021
#> 70  18-25    female           CP 4-year college State D yes 139.98052
#> 71  56-65    female           BP 4-year college State B yes  91.29599
#> 72  56-65      male Circle Party 4-year college State A yes  48.18186
#> 73  36-45      male Circle Party     associates State D yes  92.90569
#> 74  66-75      male           BP 4-year college State A yes  48.91913
#> 75  36-45 nonbinary           CP 4-year college State A yes  81.62810
#> 76  76-90      male Circle Party   some college State D yes 149.98332
#> 77  66-75    female    Box Party 4-year college State D  no  37.31033
#> 78  36-45    female           BP no high school State D yes 222.50925
#> 79  26-35      male           CP 4-year college State B  no 134.48686
#> 80  36-45      male           BP    high school State B yes  83.79017
#> 81  66-75    female Circle Party    high school State E  no 111.20261
#> 82  46-55      male Circle Party 4-year college State E yes  60.70429
#> 83  36-45      male Circle Party    high school State E yes  72.82952
#> 84  56-65    female Circle Party 4-year college State A yes 106.81903
#> 85  76-90      male Circle Party    high school State D yes  51.84883
#> 86  46-55    female           BP    high school State B yes 169.40588
#> 87  76-90      male           BP    high school State D yes  40.94947
#> 88  56-65      male           BP 4-year college State B  no  41.51055
#> 89  46-55    female           CP     associates State B yes 151.13666
#> 90  46-55    female Circle Party   some college State C  no 303.54540
#> 91  56-65    female Circle Party 4-year college State A yes 106.81903
#> 92  36-45    female           BP 4-year college State A yes 160.18389
#> 93  56-65    female Circle Party    high school State A yes 115.17939
#> 94  56-65      male Circle Party    high school State A yes  51.77488
#> 95  76-90      male Circle Party no high school State D yes  95.81371
#> 96  36-45    female    Box Party    high school State D yes  57.81996
#> 97  46-55    female    Box Party 4-year college State B yes  76.29275
#> 98  56-65      male           BP   some college State C yes  80.38327
#> 99  56-65      male           BP    high school State C yes  29.03244
#> 100 46-55    female           CP  post-graduate State A yes  75.80735
#> 101 18-25    female           BP    high school State B  no 303.54540
#> 102 56-65      male           CP    high school State C yes  19.86903
#> 103 46-55      male Circle Party    high school State B yes  96.04369
#> 104 56-65      male Circle Party    high school State B yes  56.52559
#> 105 56-65 nonbinary           BP     associates State A  no  98.41692
#> 106 56-65      male Circle Party  post-graduate State B yes  43.53646
#> 107 56-65    female Circle Party 4-year college State D  no  73.33314
#> 108 36-45      male    Box Party    high school State B  no  41.70594
#> 109 36-45      male Circle Party 4-year college State B yes  99.55316
#> 110 26-35      male           BP 4-year college State D yes 130.22139
#> 111 66-75      male Circle Party 4-year college State B yes  67.95268
#> 112 46-55      male Circle Party 4-year college State B yes  89.12213
#> 113 76-90    female           CP    high school State A yes  84.85691
#> 114 76-90      male Circle Party 4-year college State D yes  48.25025
#> 115 46-55    female           BP    high school State D yes 105.45202
#> 116 66-75      male Circle Party 4-year college State C yes  43.47107
#> 117 56-65    female Circle Party    high school State B  no 126.23352
#> 118 76-90    female Circle Party     associates State D yes 160.21446
#> 119 46-55    female Circle Party   some college State B yes 303.54540
#> 120 56-65    female           BP 4-year college State B  no  91.29599
#> 121 18-25      male Circle Party  post-graduate State B yes 162.36671
#> 122 76-90      male    Box Party    high school State D  no  21.45776
#> 123 56-65    female           BP 4-year college State A yes  83.40113
#> 124 36-45      male Circle Party    high school State A yes  97.97870
#> 125 36-45      male Circle Party     associates State C yes  93.46834
#> 126 56-65      male           CP   some college State B yes  82.46808
#> 127 46-55      male Circle Party    high school State A yes  87.71952
#> 128 56-65      male           BP 4-year college State B yes  41.51055
#> 129 56-65    female           BP    high school State A yes  89.85756
#> 130 56-65    female Circle Party    high school State B yes 126.23352
#> 131 36-45    female           BP  post-graduate State B yes 143.76533
#> 132 56-65      male Circle Party    high school State C yes  36.41759
#> 133 18-25      male Circle Party    high school State B yes 214.38457
#> 134 36-45      male Circle Party 4-year college State A  no  90.91164
#> 135 36-45    female Circle Party    high school State B yes 244.43029
#> 136 56-65    female Circle Party     associates State C  no 109.84362
#> 137 56-65      male Circle Party    high school State A yes  51.77488
#> 138 56-65      male           BP     associates State A yes  55.75918
#> 139 26-35    female Circle Party 4-year college State D yes 303.54540
#> 140 66-75      male           CP no high school State B yes  68.97143
#> 141 56-65    female           BP    high school State A  no  89.85756
#> 142 46-55      male    Box Party    high school State B yes  37.58903
#> 143 46-55    female    Box Party     associates State B  no 113.66745
#> 144 36-45      male Circle Party 4-year college State A yes  90.91164
#> 145 76-90      male           BP    high school State E yes  44.13218
#> 146 66-75    female Circle Party 4-year college State C  no  95.85780
#> 147 56-65    female Circle Party    high school State A  no 115.17939
#> 148 26-35      male           BP 4-year college State B yes 209.78773
#> 149 76-90      male    Box Party   some college State D yes  57.27717
#> 150 36-45      male           BP   some college State E yes 166.16080
#> 151 56-65 nonbinary Circle Party 4-year college State D yes  58.34690
#> 152 46-55    female           BP    high school State E yes 114.18906
#> 153 66-75      male Circle Party 4-year college State D yes  43.22281
#> 154 56-65    female           CP    high school State B yes  63.85173
#> 155 36-45 nonbinary Circle Party 4-year college State E yes 120.11171
#> 156 18-25      male Circle Party    high school State C yes 133.85761
#> 157 56-65      male Circle Party   some college State B yes 164.24949
#> 158 66-75      male Circle Party    high school State A yes  66.89884
#> 159 46-55      male           BP 4-year college State B yes  69.73440
#> 160 66-75    female           BP    high school State A yes 117.03439
#> 161 56-65    female           BP 4-year college State B  no  91.29599
#> 162 56-65    female Circle Party   some college State B  no 303.54540
#> 163 66-75    female    Box Party    high school State A  no  57.41845
#> 164 56-65      male           BP   some college State B  no 127.75276
#> 165 56-65      male Circle Party     associates State B yes  77.69032
#> 166 56-65      male Circle Party    high school State B yes  56.52559
#> 167 36-45    female Circle Party 4-year college State A  no 206.24421
#> 168 36-45    female Circle Party 4-year college State B yes 226.35162
#> 169 66-75      male           BP    high school State B yes  57.40206
#> 170 66-75    female Circle Party 4-year college State B  no 152.82248
#> 171 56-65 nonbinary    Box Party    high school State A yes  35.80269
#> 172 36-45 nonbinary           BP 4-year college State A yes 126.42809
#> 173 56-65      male Circle Party 4-year college State A yes  48.18186
#> 174 66-75      male Circle Party    high school State B yes  73.15311
#> 175 66-75      male Circle Party     associates State B  no 101.01632
#> 176 66-75    female           CP    high school State C yes  52.63444
#> 177 56-65      male Circle Party     associates State C  no  49.48173
#> 178 36-45    female           CP  post-graduate State A  no  84.60699
#> 179 66-75    female           CP   some college State B  no 244.39166
#> 180 46-55    female Circle Party     associates State A yes 277.30395
#> 181 66-75    female           BP 4-year college State C  no  74.93614
#> 182 36-45      male Circle Party 4-year college State B yes  99.55316
#> 183 36-45    female    Box Party    high school State A yes  83.81466
#> 184 26-35      male Circle Party 4-year college State C yes 168.48042
#> 185 66-75      male    Box Party    high school State C yes  19.57152
#> 186 76-90      male           BP    high school State D yes  40.94947
#> 187 66-75      male           BP    high school State B yes  57.40206
#> 188 46-55      male           CP 4-year college State A  no  41.90851
#> 189 36-45      male Circle Party    high school State C yes  67.77267
#> 190 36-45      male           BP 4-year college State A yes  71.11638
#> 191 66-75    female Circle Party    high school State A  no 150.37038
#> 192 56-65    female           BP    high school State A  no  89.85756
#> 193 76-90      male           CP 4-year college State A yes  36.13006
#> 194 18-25    female           CP 4-year college State A yes 205.64986
#> 195 36-45      male           BP   some college State C yes 154.24811
#> 196 56-65      male           BP    high school State D yes  28.87496
#> 197 56-65      male Circle Party 4-year college State B yes  52.57537
#> 198 36-45    female Circle Party    high school State E yes 164.17008
#> 199 66-75      male Circle Party 4-year college State A yes  62.16866
#> 200 56-65    female Circle Party    high school State B yes 126.23352
#> 201 56-65      male Circle Party 4-year college State D yes  33.79069
#> 202 56-65      male           BP 4-year college State A yes  38.11759
#> 203 46-55      male Circle Party    high school State A yes  87.71952
#> 204 46-55    female           BP 4-year college State C yes  98.41682
#> 205 46-55      male           BP no high school State B yes 140.38468
#> 206 18-25      male Circle Party     associates State E yes 200.61982
#> 207 46-55    female    Box Party    high school State B  no  82.17125
#> 208 56-65    female Circle Party     associates State A  no 159.97306
#> 209 56-65      male           CP 4-year college State B yes  27.78147
#> 210 56-65      male           BP  post-graduate State B yes  34.53010
#> 211 46-55    female    Box Party     associates State B  no 113.66745
#> 212 46-55      male Circle Party     associates State B yes 133.12846
#> 213 66-75      male Circle Party 4-year college State B  no  67.95268
#> 214 56-65      male           BP   some college State A yes 116.56126
#> 215 46-55      male           CP    high school State C yes  31.81421
#> 216 66-75      male Circle Party     associates State E yes  68.62816
#> 217 56-65    female           BP    high school State D yes  61.89504
#> 218 46-55      male           BP   some college State D  no 137.00532
#> 219 56-65      male Circle Party 4-year college State C yes  33.97928
#> 220 76-90    female Circle Party    high school State A yes 169.12765
#> 221 56-65      male Circle Party    high school State A yes  51.77488
#> 222 66-75    female           BP    high school State C yes  80.70434
#> 223 36-45      male Circle Party    high school State B  no 107.32280
#> 224 46-55      male Circle Party    high school State B yes  96.04369
#> 225 18-25      male           BP    high school State D  no 103.65148
#> 226 36-45      male Circle Party 4-year college State B  no  99.55316
#> 227 46-55      male           BP 4-year college State B yes  69.73440
#> 228 66-75      male    Box Party     associates State B yes  39.40406
#> 229 46-55      male           BP     associates State C  no  65.54821
#> 230 56-65    female           CP   some college State A yes 170.08484
#> 231 56-65      male           CP 4-year college State B yes  27.78147
#> 232 66-75      male           BP 4-year college State A yes  48.91913
#> 233 56-65    female    Box Party    high school State C yes  31.53064
#> 234 46-55      male           BP no high school State D yes  87.65104
#> 235 56-65    female           BP   some college State E yes 195.83920
#> 236 76-90    female           CP 4-year college State B yes  86.21157
#> 237 18-25      male Circle Party    high school State A  no 195.35929
#> 238 76-90    female Circle Party 4-year college State A yes 156.71095
#> 239 56-65    female Circle Party     associates State A  no 159.97306
#> 240 66-75      male           CP    high school State A yes  34.79565
#> 241 66-75      male Circle Party     associates State B  no 101.01632
#> 242 36-45      male    Box Party   some college State D yes  74.56054
#> 243 56-65      male    Box Party    high school State A yes  21.43077
#> 244 56-65    female Circle Party  post-graduate State B yes  96.00995
#> 245 18-25      male           BP    high school State D yes 103.65148
#> 246 46-55    female           CP 4-year college State A yes  92.22197
#> 247 66-75      male Circle Party 4-year college State B yes  67.95268
#> 248 76-90    female Circle Party 4-year college State A yes 156.71095
#> 249 56-65      male Circle Party 4-year college State B  no  52.57537
#> 250 66-75      male Circle Party   some college State A yes 195.89248
#> 251 36-45      male Circle Party 4-year college State D yes  62.60587
#> 252 56-65    female           BP   some college State C yes 181.74642
#> 253 66-75    female    Box Party 4-year college State D  no  37.31033
#> 254 56-65      male           BP no high school State A yes  74.79251
#> 255 66-75      male Circle Party     associates State A yes  92.24249
#> 256 56-65      male Circle Party   some college State B yes 164.24949
#> 257 46-55      male Circle Party no high school State B yes 180.60642
#> 258 26-35      male           CP 4-year college State D yes  84.03345
#> 259 56-65    female           BP    high school State B yes  98.39429
#> 260 46-55      male           CP no high school State D yes  57.03938
#> 261 66-75    female Circle Party 4-year college State B yes 152.82248
#> 262 36-45    female Circle Party    high school State D yes 151.47039
#> 263 46-55      male Circle Party    high school State C yes  60.81053
#> 264 66-75    female Circle Party    high school State B  no 164.92304
#> 265 46-55      male           BP    high school State D yes  47.59437
#> 266 56-65    female Circle Party    high school State A yes 115.17939
#> 267 36-45      male           BP 4-year college State A yes  71.11638
#> 268 56-65      male           BP   some college State B yes 127.75276
#> 269 26-35      male    Box Party no high school State B  no 205.78201
#> 270 66-75    female Circle Party    high school State A yes 150.37038
#> 271 56-65    female Circle Party 4-year college State C yes  73.77194
#> 272 56-65      male           BP   some college State A yes 116.56126
#> 273 56-65    female Circle Party 4-year college State D yes  73.33314
#> 274 36-45    female Circle Party 4-year college State A  no 206.24421
#> 275 76-90    female           BP 4-year college State B  no 133.65634
#> 276 46-55    female           BP 4-year college State A yes 143.13491
#> 277 66-75      male           BP   some college State A yes 152.18959
#> 278 46-55      male Circle Party 4-year college State A yes  81.42384
#> 279 36-45    female           BP 4-year college State B yes 175.71218
#> 280 36-45      male           CP   some college State B yes 158.34870
#> 281 26-35      male           BP    high school State B yes 226.51971
#> 282 66-75      male           CP   some college State B yes 107.30620
#> 283 56-65      male           BP    high school State D yes  28.87496
#> 284 66-75      male Circle Party    high school State B yes  73.15311
#> 285 56-65    female Circle Party 4-year college State C yes  73.77194
#> 286 26-35      male Circle Party 4-year college State A yes 246.37832
#> 287 36-45      male           CP 4-year college State A yes  46.55467
#> 288 36-45      male Circle Party 4-year college State D yes  62.60587
#> 289 26-35      male Circle Party 4-year college State B yes 270.47570
#> 290 56-65      male           CP   some college State B yes  82.46808
#> 291 36-45      male           BP   some college State E  no 166.16080
#> 292 46-55      male    Box Party 4-year college State A yes  32.25274
#> 293 56-65      male Circle Party   some college State D yes 102.28921
#> 294 66-75      male Circle Party    high school State D yes  46.41264
#> 295 36-45      male Circle Party 4-year college State D yes  62.60587
#> 296 36-45    female    Box Party    high school State A yes  83.81466
#> 297 56-65      male           BP    high school State A yes  40.89236
#> 298 56-65      male           BP 4-year college State B yes  41.51055
#> 299 66-75      male           CP   some college State B yes 107.30620
#> 300 66-75      male           BP   some college State D yes 103.92915
#> 301 66-75 nonbinary    Box Party     associates State D  no  43.61099
#> 302 36-45      male           CP    high school State C yes  35.22356
#> 303 18-25      male           BP    high school State B  no 166.47042
#> 304 56-65      male           BP    high school State B yes  44.56118
#> 305 56-65    female           BP 4-year college State A  no  83.40113
#> 306 66-75      male           BP   some college State B yes 166.92312
#> 307 56-65    female           BP    high school State D  no  61.89504
#> 308 56-65      male    Box Party     associates State C yes  20.59376
#> 309 46-55      male           CP    high school State A yes  44.99149
#> 310 56-65      male           BP   some college State B  no 127.75276
#> 311 56-65      male Circle Party    high school State B yes  56.52559
#> 312 76-90      male           BP 4-year college State B yes  59.71568
#> 313 66-75      male           BP    high school State A yes  52.57210
#> 314 18-25    female Circle Party no high school State D yes 303.54540
#> 315 66-75    female    Box Party    high school State E yes  43.12209
#> 316 36-45    female           BP 4-year college State B  no 175.71218
#> 317 56-65      male           BP 4-year college State A yes  38.11759
#> 318 46-55      male           BP 4-year college State B yes  69.73440
#> 319 56-65      male           BP 4-year college State B yes  41.51055
#> 320 46-55      male Circle Party 4-year college State A yes  81.42384
#> 321 76-90      male           CP 4-year college State B yes  39.32543
#> 322 56-65      male Circle Party   some college State E yes 110.75389
#> 323 46-55    female    Box Party     associates State B yes 113.66745
#> 324 36-45    female Circle Party 4-year college State D yes 140.38132
#> 325 46-55      male           BP 4-year college State D yes  44.31568
#> 326 56-65      male           BP    high school State A yes  40.89236
#> 327 66-75    female Circle Party 4-year college State B yes 152.82248
#> 328 66-75    female Circle Party    high school State A  no 150.37038
#> 329 46-55    female Circle Party 4-year college State C yes 126.26269
#> 330 18-25    female           CP    high school State C yes 151.96749
#> 331 56-65      male Circle Party 4-year college State B  no  52.57537
#> 332 18-25      male           BP   some college State D yes 303.54540
#> 333 46-55      male Circle Party 4-year college State D yes  56.20769
#> 334 56-65      male           BP     associates State D yes  38.90052
#> 335 46-55      male Circle Party 4-year college State C yes  56.53813
#> 336 56-65    female           BP 4-year college State B yes  91.29599
#> 337 56-65    female Circle Party 4-year college State B yes 117.04200
#> 338 36-45    female           BP 4-year college State B yes 175.71218
#> 339 66-75      male           BP  post-graduate State A yes  40.56042
#> 340 36-45    female    Box Party    high school State D yes  57.81996
#> 341 36-45      male           BP 4-year college State A yes  71.11638
#> 342 76-90      male Circle Party    high school State B yes  82.01581
#> 343 56-65    female           BP    high school State D yes  61.89504
#> 344 66-75    female Circle Party   some college State C yes 307.01437
#> 345 18-25      male           BP  post-graduate State C yes  79.48577
#> 346 36-45      male           BP   some college State B yes 247.41847
#> 347 46-55      male           BP    high school State C yes  47.87029
#> 348 76-90 nonbinary Circle Party 4-year college State A  no 123.70569
#> 349 56-65      male Circle Party   some college State A  no 149.75773
#> 350 66-75 nonbinary           BP  post-graduate State B yes  77.32637
#> 351 76-90      male    Box Party     associates State A yes  40.32933
#> 352 56-65      male           BP   some college State D yes  79.90290
#> 353 56-65      male           BP 4-year college State A yes  38.11759
#> 354 56-65      male           BP 4-year college State A yes  38.11759
#> 355 66-75      male    Box Party   some college State B yes  80.99781
#> 356 36-45    female    Box Party    high school State C  no  58.16059
#> 357 76-90    female           CP 4-year college State B yes  86.21157
#> 358 76-90      male Circle Party 4-year college State E  no  52.06163
#> 359 76-90      male           BP 4-year college State B yes  59.71568
#> 360 66-75      male           BP no high school State B yes 106.46819
#> 361 46-55    female           BP no high school State C  no 199.88917
#> 362 76-90    female    Box Party 4-year college State B  no  65.27463
#> 363 56-65    female           BP    high school State A yes  89.85756
#> 364 56-65      male    Box Party    high school State B yes  23.16479
#> 365 76-90      male    Box Party    high school State C yes  21.56830
#> 366 56-65    female    Box Party    high school State D yes  31.35745
#> 367 56-65    female           BP 4-year college State A yes  83.40113
#> 368 56-65      male Circle Party    high school State D yes  36.21367
#> 369 46-55      male           CP    high school State A yes  44.99149
#> 370 56-65    female Circle Party 4-year college State D yes  73.33314
#> 371 46-55      male    Box Party 4-year college State A yes  32.25274
#> 372 56-65      male           BP no high school State C yes  52.03794
#> 373 66-75    female Circle Party 4-year college State D yes  95.28012
#> 374 56-65 nonbinary           BP    high school State E yes  53.28944
#> 375 66-75      male Circle Party   some college State D yes 133.40055
#> 376 66-75      male Circle Party 4-year college State D yes  43.22281
#> 377 66-75      male    Box Party no high school State B  no  52.42446
#> 378 26-35    female Circle Party     associates State C  no 303.54540
#> 379 56-65      male    Box Party     associates State B  no  30.88999
#> 380 46-55      male Circle Party    high school State B yes  96.04369
#> 381 36-45    female    Box Party    high school State A yes  83.81466
#> 382 18-25      male           BP 4-year college State C yes  96.74112
#> 383 56-65    female    Box Party   some college State A  no 127.79074
#> 384 26-35      male           BP   some college State B  no 303.54540
#> 385 66-75 nonbinary    Box Party 4-year college State A  no  42.72243
#> 386 36-45      male           CP 4-year college State D yes  32.69338
#> 387 66-75      male           BP no high school State A yes  97.20137
#> 388 36-45      male Circle Party 4-year college State B  no  99.55316
#> 389 56-65    female           BP 4-year college State A  no  83.40113
#> 390 46-55      male Circle Party   some college State A  no 259.40470
#> 391 76-90    female           CP   some college State D yes 170.34189
#> 392 66-75    female    Box Party 4-year college State B  no  58.31348
#> 393 18-25      male Circle Party 4-year college State B yes 198.56502
#> 394 66-75      male           CP 4-year college State D yes  23.20153
#> 395 56-65      male Circle Party no high school State D yes  65.81593
#> 396 66-75      male           BP 4-year college State B yes  53.38594
#> 397 46-55      male Circle Party 4-year college State C yes  56.53813
#> 398 56-65      male           CP 4-year college State C yes  18.67499
#> 399 66-75      male Circle Party  post-graduate State B yes  56.05303
#> 400 46-55      male           BP 4-year college State A yes  63.78927
#> 401 56-65    female           CP 4-year college State A yes  54.34449
#> 402 46-55      male           BP   some college State A yes 201.23797
#> 403 36-45      male    Box Party    high school State A yes  38.29531
#> 404 66-75      male           BP 4-year college State C yes  34.47961
#> 405 56-65    female    Box Party    high school State B yes  48.60842
#> 406 26-35    female           CP    high school State E yes 222.72455
#> 407 66-75 nonbinary Circle Party 4-year college State D yes  75.55086
#> 408 46-55    female           BP 4-year college State D yes  97.82305
#> 409 46-55      male    Box Party  post-graduate State B yes  29.28174
#> 410 66-75      male           BP   some college State E yes 112.53506
#> 411 66-75      male           BP 4-year college State A yes  48.91913
#> 412 18-25 nonbinary           CP     associates State B yes 267.62036
#> 413 66-75    female           BP  post-graduate State A yes  89.08519
#> 414 56-65      male           BP 4-year college State B yes  41.51055
#> 415 66-75    female Circle Party    high school State A yes 150.37038
#> 416 56-65    female Circle Party  post-graduate State A yes  87.68882
#> 417 56-65      male           BP 4-year college State B yes  41.51055
#> 418 66-75      male    Box Party 4-year college State B  no  27.33572
#> 419 36-45    female Circle Party  post-graduate State B yes 184.98399
#> 420 56-65    female           BP 4-year college State D yes  57.54109
#> 421 36-45    female    Box Party     associates State A  no 115.97291
#> 422 56-65    female           BP 4-year college State A yes  83.40113
#> 423 76-90      male Circle Party    high school State A yes  74.96012
#> 424 56-65    female Circle Party   some college State A yes 303.54540
#> 425 56-65    female    Box Party 4-year college State B  no  45.25348
#> 426 46-55      male Circle Party   some college State C yes 177.32044
#> 427 26-35      male           BP 4-year college State C yes 131.02017
#> 428 56-65      male Circle Party 4-year college State A yes  48.18186
#> 429 66-75    female    Box Party  post-graduate State B  no  48.20707
#> 430 56-65      male Circle Party    high school State E yes  38.98858
#> 431 56-65    female           BP     associates State B yes 136.42601
#> 432 46-55    female Circle Party    high school State B yes 218.18566
#> 433 56-65      male Circle Party 4-year college State A  no  48.18186
#> 434 56-65    female           CP 4-year college State B yes  59.35066
#> 435 46-55    female Circle Party 4-year college State B yes 202.08031
#> 436 66-75      male           BP 4-year college State C yes  34.47961
#> 437 76-90    female Circle Party    high school State A yes 169.12765
#> 438 56-65    female           BP   some college State D  no 180.62868
#> 439 56-65      male Circle Party   some college State C yes 102.91123
#> 440 36-45    female           BP 4-year college State B  no 175.71218
#> 441 56-65    female           BP 4-year college State B  no  91.29599
#> 442 46-55      male           CP   some college State A yes 129.06541
#> 443 56-65      male Circle Party    high school State D yes  36.21367
#> 444 76-90      male Circle Party   some college State C yes 150.90715
#> 445 56-65 nonbinary Circle Party 4-year college State B yes  92.60971
#> 446 56-65      male Circle Party    high school State D yes  36.21367
#> 447 66-75    female Circle Party 4-year college State A yes 139.36402
#> 448 46-55      male           BP    high school State C yes  47.87029
#> 449 36-45      male    Box Party 4-year college State D yes  25.38411
#> 450 76-90      male           BP     associates State A yes  80.87756
#> 451 56-65    female Circle Party 4-year college State B  no 117.04200
#> 452 56-65 nonbinary Circle Party    high school State A  no  91.14964
#> 453 46-55      male Circle Party 4-year college State A yes  81.42384
#> 454 46-55    female Circle Party    high school State B  no 218.18566
#> 455 56-65    female           BP 4-year college State D yes  57.54109
#> 456 56-65    female           BP 4-year college State A  no  83.40113
#> 457 56-65      male Circle Party    high school State D yes  36.21367
#> 458 36-45      male           BP  post-graduate State B  no  64.06020
#> 459 66-75    female           BP    high school State A yes 117.03439
#> 460 56-65    female           BP 4-year college State C yes  57.87996
#> 461 18-25      male           BP     associates State C yes 144.68578
#> 462 66-75    female           CP 4-year college State E  no  52.54349
#> 463 66-75      male Circle Party 4-year college State A yes  62.16866
#> 464 56-65      male           BP     associates State E yes  41.90680
#> 465 56-65      male Circle Party     associates State A yes  71.02577
#> 466 56-65      male Circle Party    high school State A yes  51.77488
#> 467 76-90    female Circle Party   some college State C yes 303.54540
#> 468 66-75      male Circle Party    high school State A yes  66.89884
#> 469 66-75      male           CP 4-year college State A yes  32.47928
#> 470 26-35      male Circle Party    high school State D  no 180.73559
#> 471 56-65    female           BP    high school State A yes  89.85756
#> 472 36-45    female           BP 4-year college State A yes 160.18389
#> 473 46-55      male Circle Party     associates State A yes 121.45085
#> 474 36-45      male           BP    high school State B yes  83.79017
#> 475 56-65      male           BP    high school State D yes  28.87496
#> 476 56-65      male Circle Party    high school State C yes  36.41759
#> 477 56-65      male    Box Party  post-graduate State C  no  12.89883
#> 478 36-45      male    Box Party   some college State D yes  74.56054
#> 479 76-90      male           BP  post-graduate State A yes  45.24671
#> 480 66-75    female    Box Party  post-graduate State B  no  48.20707
#> 481 56-65      male           BP 4-year college State B yes  41.51055
#> 482 18-25      male    Box Party    high school State D yes  51.09317
#> 483 56-65      male           BP    high school State D yes  28.87496
#> 484 56-65    female Circle Party   some college State D  no 232.71795
#> 485 46-55    female           CP     associates State E yes 102.01817
#> 486 56-65      male Circle Party    high school State B yes  56.52559
#> 487 46-55      male           CP   some college State C yes  88.86890
#> 488 56-65    female           BP 4-year college State C yes  57.87996
#> 489 56-65      male           BP     associates State A yes  55.75918
#> 490 76-90    female           BP   some college State A  no 303.54540
#> 491 76-90    female Circle Party 4-year college State A  no 156.71095
#> 492 46-55    female           CP    high school State B  no 108.88053
#> 493 36-45      male           BP 4-year college State B yes  77.78994
#> 494 36-45      male    Box Party   some college State A yes 108.63912
#> 495 66-75      male           BP 4-year college State A yes  48.91913
#> 496 36-45      male           BP    high school State C yes  53.24692
#> 497 66-75      male Circle Party    high school State C yes  46.68109
#> 498 56-65      male    Box Party    high school State A yes  21.43077
#> 499 36-45      male Circle Party 4-year college State D yes  62.60587
#> 500 26-35      male Circle Party    high school State B  no 292.14176